+++ examples/qnanopainter_vs_qpainter_demo/src/demoqnanoitempainter.cpp
@@ -44,39 +44,39 @@ void DemoQNanoItemPainter::paint(QNanoPainter *painter)
 {
     m_painter = painter;
 
-    qreal w = width();
-    qreal h = height();
-    qreal t = m_animationTime;
+    float w = width();
+    float h = height();
+    float t = m_animationTime;
 
     // These painting commands are identical with both renderers
     for (int i=0 ; i < m_testCount ; i++) {
         // Paint ruler
         if (m_enabledTests & 1) {
-            drawRuler(0, h*0.02, w, h*0.05, t);
+            drawRuler(0, h*0.02f, w, h*0.05f, t);
         }
         // Paint circles
         if (m_enabledTests & 2) {
-            drawGraphCircles(w*0.15, h*0.1, w*0.7, w*0.7, 10, t*2);
-            drawGraphCircles(w*0.05, h*0.55, w*0.25, w*0.25, 8, t*3);
-            drawGraphCircles(w*0.70, h*0.55, w*0.25, w*0.25, 3, t);
+            drawGraphCircles(w*0.15f, h*0.1f, w*0.7f, w*0.7f, 10, t*2);
+            drawGraphCircles(w*0.05f, h*0.55f, w*0.25f, w*0.25f, 8, t*3);
+            drawGraphCircles(w*0.70f, h*0.55f, w*0.25f, w*0.25f, 3, t);
         }
         // Paint lines
         if (m_enabledTests & 4) {
             drawGraphLine(0, h, w, -h, 4, t);
-            drawGraphLine(0, h, w, -h*0.8, 6,  t+10);
-            drawGraphLine(0, h, w, -h*0.6, 12, t/2);
+            drawGraphLine(0, h, w, -h*0.8f, 6,  t+10);
+            drawGraphLine(0, h, w, -h*0.6f, 12, t/2);
         }
         // Paint bars
         if (m_enabledTests & 8) {
-            drawGraphBars(0, h, w, -h*0.8, 6, t*3);
-            drawGraphBars(0, h, w, -h*0.4, 10, t+2);
-            drawGraphBars(0, h, w, -h*0.3, 20, t*2+2);
-            drawGraphBars(0, h, w, -h*0.2, 40, t*3+2);
+            drawGraphBars(0, h, w, -h*0.8f, 6, t*3);
+            drawGraphBars(0, h, w, -h*0.4f, 10, t+2);
+            drawGraphBars(0, h, w, -h*0.3f, 20, t*2+2);
+            drawGraphBars(0, h, w, -h*0.2f, 40, t*3+2);
         }
         // Paint icons
         int icons = 50;
         if (m_enabledTests & 16) {
-            drawIcons(0, h*0.2, w, h*0.2, icons, t);
+            drawIcons(0, h*0.2f, w, h*0.2f, icons, t);
         }
 
         // Paint flower
@@ -142,13 +142,13 @@ void DemoQNanoItemPainter::drawGraphBars(float x, float y, float w, float h, int
     QVarLengthArray<float, 1024> samples(items);
     QVarLengthArray<float, 1024> sx(items); QVarLengthArray<float, 1024> sy(items);
     float dx = w/items;
-    float barWidth = dx * 0.8;
+    float barWidth = dx * 0.8f;
     float margin = dx - barWidth;
     int i;
 
     // Generate positions
     for (i = 0; i<items; i++) {
-        samples[i] = 0.5 + sinf(i*0.1+t)*0.5;
+        samples[i] = 0.5f + sinf(i*0.1f+t)*0.5f;
     }
     for (i = 0; i < items; i++) {
         sx[i] = x+i*dx + margin/2;
@@ -158,9 +158,9 @@ void DemoQNanoItemPainter::drawGraphBars(float x, float y, float w, float h, int
     // Draw graph bars
     m_painter->beginPath();
     for (i = 0; i < items; i++) {
-        m_painter->rect((int)sx[i]+0.5, (int)y+1.5, (int)barWidth, (int)sy[i]);
+        m_painter->rect((int)sx[i]+0.5f, (int)y+1.5f, (int)barWidth, (int)sy[i]);
     }
-    qreal lineWidth = 0.5 + w * 0.002;
+    float lineWidth = 0.5f + w * 0.002f;
     m_painter->setLineWidth(lineWidth);
     m_painter->setLineJoin(QNanoPainter::JOIN_MITER);
     m_painter->setFillStyle(m_color3);
@@ -171,19 +171,19 @@ void DemoQNanoItemPainter::drawGraphBars(float x, float y, float w, float h, int
 
 void DemoQNanoItemPainter::drawGraphCircles(float x, float y, float w, float h, int items, float t)
 {
-    qreal barWidth = 0.3 * w/items;
-    qreal lineMargin = 0.2 * barWidth;
-    qreal showAnimationProgress = 0.1 + 0.4*sinf(t*0.8)+0.5;
-    qreal lineWidth = barWidth*showAnimationProgress;
+    float barWidth = 0.3f * w/items;
+    float lineMargin = 0.2f * barWidth;
+    float showAnimationProgress = 0.1f + 0.4f*sinf(t*0.8f)+0.5f;
+    float lineWidth = barWidth*showAnimationProgress;
 
-    qreal cx = x+w/2;
-    qreal cy = y+h/2;
-    qreal radius1 = w/2 - lineWidth;
+    float cx = x+w/2;
+    float cy = y+h/2;
+    float radius1 = w/2 - lineWidth;
     int i;
 
     // Setup values
-    qreal a1 = -M_PI/2;
-    QVarLengthArray<qreal, 1024> a0(items);
+    float a1 = -M_PI/2;
+    QVarLengthArray<float, 1024> a0(items);
     for (i=0; i<items; i++) {
         a0[i] = -M_PI/2 + 2*M_PI*(((float)items-i)/items)*showAnimationProgress;
     }
@@ -193,7 +193,7 @@ void DemoQNanoItemPainter::drawGraphCircles(float x, float y, float w, float h, 
     m_painter->setLineCap(QNanoPainter::CAP_ROUND);
 
     // Draw cicle backgrounds
-    qreal r = radius1;
+    float r = radius1;
     QNanoColor c_background(215,215,215,50);
     m_painter->setStrokeStyle(c_background);
     for (int i=0 ; i<items ; i++) {
@@ -219,16 +219,16 @@ void DemoQNanoItemPainter::drawGraphCircles(float x, float y, float w, float h, 
 void DemoQNanoItemPainter::drawIcons(float x, float y, float w, float h, int items, float t)
 {
     // Note: Adjust font size to match QPainter sizing
-    qreal fontSize = w/22.0 * 1.32;
+    float fontSize = w/22.0f * 1.32f;
     m_testFont.setPixelSize(fontSize);
     m_painter->setFont(m_testFont);
     m_painter->setFillStyle("#FFFFFF");
     m_painter->setTextAlign(QNanoPainter::ALIGN_CENTER);
     m_painter->setTextBaseline(QNanoPainter::BASELINE_MIDDLE);
-    qreal size = w/12;
+    float size = w/12;
     for (int i=0 ; i<items ; i++) {
-        qreal xp = x + (w-size)/items*i;
-        qreal yp = y + h*0.5 + h * sinf((i+1) * t * 0.1) * 0.5;
+        float xp = x + (w-size)/items*i;
+        float yp = y + h*0.5f + h * sinf((i+1) * t * 0.1) * 0.5f;
         m_painter->drawImage(m_circleImage, QRectF(xp, yp, size, size));
         m_painter->fillText(QString::number(i+1), QRectF(xp, yp+size/2, size, size));
     }
@@ -236,12 +236,12 @@ void DemoQNanoItemPainter::drawIcons(float x, float y, float w, float h, int ite
 
 void DemoQNanoItemPainter::drawRuler(float x, float y, float w, float h, float t)
 {
-    float posX = x + w*0.05;
-    double space = w*0.03 + sinf(t)*w*0.02;
+    float posX = x + w*0.05f;
+    double space = w*0.03f + sinf(t)*w*0.02f;
     m_painter->setTextAlign(QNanoPainter::ALIGN_CENTER);
     m_painter->setTextBaseline(QNanoPainter::BASELINE_MIDDLE);
     // Note: Adjust font size to match QPainter sizing
-    qreal fontSize = w/35.0 * 1.32;
+    float fontSize = w/35.0f * 1.32f;
     m_testFont.setPixelSize(fontSize);
     m_painter->setFont(m_testFont);
     m_painter->setStrokeStyle("#E0E0E0");
